move history code into HistoryService and out of LocalWebApp routes
