+++ backend/FwLite/LcmCrdt/LcmCrdt.csproj
@@ -13,6 +13,7 @@
     </ItemGroup>
 
     <ItemGroup>
+        <PackageReference Include="Humanizer.Core" Version="2.14.1"/>
         <PackageReference Include="linq2db.AspNet" Version="5.4.1"/>
         <PackageReference Include="linq2db.EntityFrameworkCore" Version="8.1.0"/>
         <PackageReference Include="linq2db.SQLite" Version="5.4.1" />
