+++ backend/FwLite/LocalWebApp/Routes/ActivityRoutes.cs
@@ -25,38 +25,7 @@ public static class ActivityRoutes
             });
             return operation;
         });
-        group.MapGet("/",
-            (ICrdtDbContext dbcontext) =>
-            {
-                return dbcontext.Commits
-                    .DefaultOrderDescending()
-                    .Take(20)
-                    .Select(c => new Activity(c.Id, c.HybridDateTime.DateTime, c.ChangeEntities))
-                    .AsAsyncEnumerable();
-            });
+        group.MapGet("/", (HistoryService historyService) => historyService.ProjectActivity());
         return group;
-    }
-
-    private static string ChangeName(List<ChangeEntity<IChange>> changeEntities)
-    {
-        return changeEntities switch
-        {
-            { Count: 0 } => "No changes",
-            { Count: 1 } => changeEntities[0].Change switch
-            {
-                //todo call JsonPatchChange.Summarize() instead of this
-                IChange change when change.GetType().Name.StartsWith("JsonPatchChange") => "Change " + change.EntityType.Name,
-                IChange change => change.GetType().Name.Humanize()
-            },
-            { Count: var count } => $"{count} changes"
-        };
-    }
-
-    public record Activity(
-        Guid CommitId,
-        DateTimeOffset Timestamp,
-        List<ChangeEntity<IChange>> Changes)
-    {
-        public string ChangeName => ChangeName(Changes);
     }
 }
